requirements.txt
Flask
python-docx
PyMuPDF
nltk
python-dotenv
numpy
requests
gunicorn app:app